chore(constants): 카테고리 매핑 데이터를 constants 폴더로 이동
